blueprints/gke/hpc/bundle/app/images/video-downloader-creator/requirements.txt
internetarchive==3.3.0
PyYAML==6.0
kubernetes==26.1.0